production/requirements/requirements.txt
# packaging
setuptools==65.5.0
wheel==0.37.1

# production requirements
d2l==1.0.0-beta0
torchvision==0.14.1

# testing requirements
pytest==7.1.2